# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for Stitchfolio.",
        "title": "Stitchfolio-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/measurement": {
            "get": {
                "description": "Get all active measurements grouped by PersonId with DressTypes as CSV. Searches by Customer Name.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Get all active measurements grouped by PersonId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by Customer Name (returns all Persons of that customer)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.MeasurementBrowse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves an instance of Measurement",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Save Measurement",
                "parameters": [
                    {
                        "description": "measurement",
                        "name": "measurement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Measurement"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
...
        "/measurement/bulk": {
            "put": {
                "description": "Updates an array of measurements by their IDs",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Bulk Update Measurements",
                "parameters": [
                    {
                        "description": "Array of measurements to update",
                        "name": "measurements",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/requestModel.Measurement"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves multiple measurements for multiple persons in bulk",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Save Bulk Measurements",
                "parameters": [
                    {
                        "description": "Array of bulk measurement requests",
                        "name": "measurements",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/requestModel.BulkMeasurementRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
        "/measurement/{id}": {
            "get": {
                "description": "Get an instance of Measurement",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Get a specific Measurement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Measurement id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Measurement"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a single measurement by its ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Update Measurement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Measurement id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Measurement to update",
                        "name": "measurement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Measurement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an instance of Measurement",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Measurement"
                ],
                "summary": "Delete Measurement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "measurement id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestModel.BulkMeasurementItem": {
            "type": "object",
            "properties": {
                "dressTypeId": {
                    "type": "integer"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "requestModel.BulkMeasurementRequest": {
            "type": "object",
            "properties": {
                "measurements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requestModel.BulkMeasurementItem"
                    }
                },
                "personId": {
                    "type": "integer"
                }
            }
        },
...
        "requestModel.Measurement": {
            "type": "object",
            "properties": {
                "dressTypeId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "personId": {
                    "type": "integer"
                },
                "takenById": {
                    "type": "integer"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "responseModel.AuditFields": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "email": {
                    "type": "string"
                },
                "enquiries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseModel.Enquiry"
                    }
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseModel.Order"
                    }
                },
                "persons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseModel.Person"
                    }
                },
                "phoneNumber": {
                    "type": "string"
                },
                "whatsappNumber": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.DataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
        "responseModel.DressType": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "measurements": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "responseModel.Enquiry": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "customer": {
                    "$ref": "#/definitions/responseModel.Customer"
                },
                "customerId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "notes": {
                    "type": "string"
                },
                "referredBy": {
                    "type": "string"
                },
                "referrerPhoneNumber": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Measurement": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "dressType": {
                    "$ref": "#/definitions/responseModel.DressType"
                },
                "dressTypeId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "person": {
                    "$ref": "#/definitions/responseModel.Person"
                },
                "personId": {
                    "type": "integer"
                },
                "personName": {
                    "type": "string"
                },
                "takenBy": {
                    "description": "first_name + last_name",
                    "type": "string"
                },
                "takenById": {
                    "type": "integer"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "responseModel.MeasurementBrowse": {
            "type": "object",
            "properties": {
                "customerId": {
                    "type": "integer"
                },
                "dressTypes": {
                    "description": "CSV of DressType names",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "personId": {
                    "type": "integer"
                },
                "personName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Order": {
            "type": "object",
            "properties": {
                "additionalCharges": {
                    "type": "number"
                },
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "customer": {
                    "$ref": "#/definitions/responseModel.Customer"
                },
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "description": "first_name + last_name",
                    "type": "string"
                },
                "deliveredDate": {
                    "type": "string"
                },
                "expectedDeliveryDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "notes": {
                    "type": "string"
                },
                "orderItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseModel.OrderItem"
                    }
                },
                "orderQuantity": {
                    "description": "sum of quantity from order items",
                    "type": "integer"
                },
                "orderTakenBy": {
                    "description": "first_name + last_name",
                    "type": "string"
                },
                "orderTakenById": {
                    "type": "integer"
                },
                "orderValue": {
                    "description": "sum of total from order items",
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.OrderItem": {
            "type": "object",
            "properties": {
                "additionalCharges": {
                    "type": "number"
                },
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "deliveredDate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expectedDeliveryDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "measurement": {
                    "$ref": "#/definitions/responseModel.Measurement"
                },
                "measurementId": {
                    "type": "integer"
                },
                "order": {
                    "$ref": "#/definitions/responseModel.Order"
                },
                "orderId": {
                    "type": "integer"
                },
                "person": {
                    "$ref": "#/definitions/responseModel.Person"
                },
                "personId": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "responseModel.Person": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "customer": {
                    "$ref": "#/definitions/responseModel.Customer"
                },
                "customerId": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "measurements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseModel.Measurement"
                    }
                }
            }
        },
...
        "responseModel.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}